Added new module 'core/mathematics'
